Add a method to call for a camera shake effect
This is just for if we want to do that. Nothign is changed yet, just put
the method in place. We currently cannot shake the camer since the
camera has a follow script that controls its transform. The shake effect
also alters the transform, so the followscript will override any shake
effect. When we do somethign different with the camera follow
(cinimachine or whatever), maybe then we can have a ScreenShake. :)